// apps/backend/src/types/Status.enum.ts
export enum Job_Status {
  PUBLISHED = 'published',
  CONFIRMED = 'confirmed',
  FINISHED = 'finished',
  CLOSED = 'closed',
  CANCELLED = 'cancelled',
  DELETED = 'deleted',
}

export enum Offer_Status {
  PENDING = 'pending',
  ACCEPTED = 'accepted',
  REJECTED = 'rejected',
  CANCELLED = 'cancelled',
}